[Tests] Split yaspgrid tests into multiple ones in a separate directory, reduced run time for 3d case
